allow for configuring oauth scopes
